Geospatial Project Learning
# US Wildfires - geospatial visualizations with Altair

We will demonstrate how to create **interactive visualizations of geospatial data** in Python. The landscape of Python plotting libraries is truly vast, but here we willl focus on [Altair](https://altair-viz.github.io/), an easy-to-use Python package that enables a broad range of interactivity in just a few lines of code.

Let's first get the data!

## US Wildfires data set

The data set we will be working with contains over 20 years of geo-referenced US wildfire records information for fires that were larger than 100 acres across. The data spans from 1992 to 2015 and contains various information about each fire incident including the date of the fire, latitude, longitude, fire size, etc.